Feature/create workspaces

+++ tests/integration_tests/adapters/mysql/generic_mysql_test.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 from uuid import uuid4
 
 from c4maker_server.domain.exceptions.duplicate_entity_exception import DuplicateEntityException
-from tests.integration_tests.base_integ_test import BaseIntegTest
+from tests.integration_tests.default_values import DefaultValues
 
 
 class GenericMySQLTest(metaclass=abc.ABCMeta):
@@ -41,14 +41,14 @@ class GenericMySQLTest(metaclass=abc.ABCMeta):
 
         for key1, value1 in dict1.items():
             value2 = dict2.get(key1)
-            self.get_test_case().assertEqual(value1, value2, f"value expected for {key1} is value1 but has {value2}")
+            self.get_test_case().assertEqual(value1, value2, f"value expected for {key1} is {value1} but has {value2}")
 
     def test_find_entity(self):
         # given
         default_entity = self.get_default_entity()
 
         # when
-        persisted_entity = self.get_repository().find_by_id(BaseIntegTest.DEFAULT_ID)
+        persisted_entity = self.get_repository().find_by_id(DefaultValues.DEFAULT_ID)
 
         # then
         self.compare_obj_properties(default_entity, persisted_entity)
@@ -87,7 +87,10 @@ class GenericMySQLTest(metaclass=abc.ABCMeta):
                                          str(ex_context.exception))
 
     def test_update_entity(self):
-        entity = self.get_repository().find_by_id(BaseIntegTest.DEFAULT_ID)
+        if not self.get_updated_default_entity():
+            return
+
+        entity = self.get_repository().find_by_id(DefaultValues.DEFAULT_ID)
         # ensuring that entity exists and match with default entity
         self.compare_obj_properties(self.get_default_entity(), entity)
 
@@ -96,24 +99,30 @@ class GenericMySQLTest(metaclass=abc.ABCMeta):
         self.get_repository().update(updated_entity)
 
         # when
-        persisted_entity = self.get_repository().find_by_id(BaseIntegTest.DEFAULT_ID)
+        persisted_entity = self.get_repository().find_by_id(DefaultValues.DEFAULT_ID)
 
         # then
         self.compare_obj_properties(updated_entity, persisted_entity)
 
     def test_delete_entity(self):
+        if not self.get_updated_default_entity():
+            return
+
         # given
-        entity = self.get_repository().find_by_id(BaseIntegTest.DEFAULT_ID)
+        entity = self.get_repository().find_by_id(DefaultValues.DEFAULT_ID)
         # ensuring that entity exists
         self.get_test_case().assertIsNotNone(entity)
 
         # when
-        self.get_repository().delete(BaseIntegTest.DEFAULT_ID)
+        self.get_repository().delete(DefaultValues.DEFAULT_ID)
 
         # then
-        entity = self.get_repository().find_by_id(BaseIntegTest.DEFAULT_ID)
+        entity = self.get_repository().find_by_id(DefaultValues.DEFAULT_ID)
         self.get_test_case().assertIsNone(entity)
 
     def test_delete_non_exists_entity(self):
+        if not self.get_updated_default_entity():
+            return
+
         # nothing is expected. Just to test behaviour
         self.get_repository().delete(uuid4())
